+++ lib/subvisual/string_utils.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "subvisual/string_utils/replace_at"
+
+module Subvisual
+  module StringUtils
+    extend ReplaceAt
+  end
+end