+++ lib/subvisual/string_utils/replace_at.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Subvisual
+  module StringUtils
+    module ReplaceAt
+      def replace_at(string, index, replacement)
+        string.chars.tap do |characters|
+          characters[index] = replacement
+        end.join
+      end
+    end
+  end
+end